pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = "LinkedIn job application agent with human-in-the-loop approval"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.5",
    "anthropic>=0.96.0",
    "apscheduler>=3.11.2",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "playwright>=1.58.0",
    "pydantic>=2.13.1",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.2.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "python-telegram-bot>=22.7",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "weasyprint>=68.1",
]

[project.scripts]
pa = "personal_assistant.cli:main"
